+++ app/src/main/java/woowacourse/paint/drawing/Drawing.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.drawing
 
 import android.graphics.Paint
 import android.graphics.Path
