+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="viewModel"
-            type="woowacourse.paint.MainViewModel" />
+            type="woowacourse.paint.ui.viewmodel.MainViewModel" />
         <variable
             name="defaultPaintWidth"
             type="float" />
@@ -18,7 +18,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <woowacourse.paint.Canvas
+        <woowacourse.paint.ui.custom.Canvas
             android:id="@+id/main_canvas"
             android:layout_width="match_parent"
             android:layout_height="0dp"
